requirements.txt
Django==5.0.0     
djangorestframework==3.15.0 
djangorestframework-simplejwt==5.3.1 
langchain
langchain-openai